+++ CHANGELOG.rst
@@ -1,6 +1,23 @@
 Changelog
 =========
 
+v0.1.2 (2020-01-06)
+-------------------
+
+* Add initial Sphinx documentation skeleton to project
+
+v0.1.1 (2020-01-04)
+-------------------
+
+* Simplify packaging and remove tox.ini and MANIFEST.in from template
+* Add travis.yml, .editorconfig, and Pipfile to template
+* Add initial readme to project
+
+v0.1.0 (2020-01-02)
+-------------------
+
+* Add basic python packaging files to template.
+
 v0.0.0 (2020-01-01)
 -------------------